--- Kernel/Src/Device/AHCI/AHCI.cpp
@@ -1,282 +1,0 @@
-/*
- *  Copyright 2025 Ewogijk
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-
-#include <Device/AHCI/AHCI.h>
-
-#include <Memory/MemoryModule.h>
-#include <Memory/Paging.h>
-
-namespace Rune::Device {
-    const SharedPointer<Logger> LOGGER = LogContext::instance().get_logger("Device.AHCI");
-
-    auto AHCIDriver::alloc_system_memory(U32 ct_count) -> SystemMemory* {
-        auto* sys_mem = reinterpret_cast<SystemMemory*>(_heap->allocate(sizeof(SystemMemory)));
-
-        void* clb = _heap->allocate_dma(SystemMemory::COMMAND_LIST_SIZE * sizeof(CommandHeader));
-        if (clb == nullptr) {
-            LOGGER->error("Failed to allocate command list.");
-            return nullptr;
-        }
-        sys_mem->CL = reinterpret_cast<CommandHeader*>(clb);
-
-        void* fb = _heap->allocate_dma(sizeof(ReceivedFIS));
-        if (fb == nullptr) {
-            LOGGER->error("Failed to allocate received FIS...");
-            _heap->free(clb);
-            return nullptr;
-        }
-        sys_mem->RFIS = reinterpret_cast<ReceivedFIS*>(fb);
-
-        void* ct = _heap->allocate_dma(ct_count * sizeof(CommandTable));
-        if (ct == nullptr) {
-            LOGGER->error("Failed to allocate command tables...");
-            _heap->free(clb);
-            _heap->free(fb);
-            return nullptr;
-        }
-        sys_mem->CT           = reinterpret_cast<CommandTable*>(ct);
-        sys_mem->CommandSlots = (int8_t) ct_count;
-
-        for (U32 j = 0; j < ct_count; j++) {
-            PhysicalAddr p_ctba{0};
-            if (!Memory::virtual_to_physical_address(
-                    memory_pointer_to_addr(&reinterpret_cast<CommandTable*>(ct)[j]),
-                    p_ctba)) {
-                LOGGER->error("Failed hook command table {} into system memory!", j);
-                _heap->free(sys_mem->CL);
-                _heap->free(sys_mem->RFIS);
-                _heap->free(sys_mem->CT);
-                return nullptr;
-            }
-
-            sys_mem->CL[j].CTBA.AsUInt32 = (U32) p_ctba;
-            if (sys_mem->CL[j].CTBA.Reserved != 0) {
-                LOGGER->error("Command table base address is not 128 byte aligned!");
-                _heap->free(sys_mem->CL);
-                _heap->free(sys_mem->RFIS);
-                _heap->free(sys_mem->CT);
-                return nullptr;
-            }
-#ifdef IS_64_BIT
-            if (_hba->CAP.S64A) sys_mem->CL[j].CTBAU = (U32) (p_ctba >> 32);
-#endif
-            sys_mem->CL[j].PRDTL = 1;
-        }
-
-        return sys_mem;
-    }
-
-    auto AHCIDriver::resolve_logical_drive(U8 logicalDrive) -> LogicalDrive {
-        if (logicalDrive > LOGICAL_DRIVE_LIMIT) {
-            LOGGER->warn("Invalid logical drive ID: {}", logicalDrive);
-            return {};
-        }
-
-        auto ld = _logical_drive_table[logicalDrive];
-        if (ld.port_index == LogicalDrive::INVALID_PORT) {
-            LOGGER->warn("Logical drive {} not found.", logicalDrive);
-            return {};
-        }
-        return ld;
-    }
-
-    const PCIDeviceID AHCIDriver::ID_AHCI(0x1, 0x6, 0x1);
-
-    AHCIDriver::AHCIDriver(DriverHandle handle, Memory::SlabAllocator* heap, CPU::Timer* timer)
-        : Driver(handle, "AHCI"),
-          _heap(heap),
-          _timer(timer) {}
-
-    // ========================================================================================== //
-    // Driver API
-    // ========================================================================================== //
-
-    auto AHCIDriver::get_target_device_ID() const -> const DeviceID* { return &ID_AHCI; }
-
-    auto AHCIDriver::start(void* context) -> bool {
-        auto* pci_type0_header = static_cast<PCIConfigurationSpaceHeaderType0*>(context);
-        _hba                   = reinterpret_cast<HBAMemory*>(
-            Memory::physical_to_virtual_address(pci_type0_header->bar_5));
-        _hba->GHC.AE = 1;
-        return true;
-    }
-
-    auto AHCIDriver::stop() -> bool { return _hba == nullptr; }
-
-    auto AHCIDriver::handle_request(IORequest request) -> IOResponse {
-        return {.m_status = IORequestStatus::FAILED, .m_data = nullptr};
-    }
-
-    void AHCIDriver::discover_devices(DeviceHandle                 bus_device,
-                                      const DeviceMapper&          device_mapper,
-                                      HandleCounter<DeviceHandle>& dev_handle_counter) {
-        U32  pi                 = _hba->PI;
-        U32  command_slots      = _hba->CAP.NCS;
-        bool s64_a              = _hba->CAP.S64A;
-        U8   c_logical_drive_id = 0;
-        for (size_t i = 0; i < PORT_LIMIT; i++) {
-            if (_logical_drive_count == LOGICAL_DRIVE_LIMIT) {
-                LOGGER->warn("Limit of 255 logical drives reached. Stopping port scan... ");
-                break;
-            }
-
-            if (!bit_check(pi, i)) continue;
-
-            LOGGER->debug("------------------------------------- Scanning Port {} "
-                          "-------------------------------------",
-                          i);
-            // NOLINTBEGIN iterating -> index is in bounds
-            if (!_port_engine[i].scan_device(&_hba->Port[i])) continue;
-            // NOLINTEND
-            if (!_port_engine[i].stop()) {
-                LOGGER->error("Stopping the port failed. Trying port reset...");
-                _port_engine[i].reset();
-            }
-
-            SystemMemory* system_memory = alloc_system_memory(command_slots);
-            if (!_port_engine[i].start(system_memory, s64_a, _heap, _timer)) {
-                LOGGER->error("Failed to start port {}. Freeing allocated system memory...", i);
-                _heap->free(system_memory->CL);
-                _heap->free(system_memory->CT);
-                _heap->free(system_memory->RFIS);
-            }
-
-            LOGGER->debug("Detected logical drives:");
-            auto hd_info = _port_engine[i].get_hard_drive_info();
-            auto pt      = hd_info.partition_table;
-            for (size_t j = 0; j < pt.size(); j++) {
-                auto* partition = pt[j];
-                auto  handle    = dev_handle_counter.acquire();
-                auto  device    = SharedPointer<Device>(
-                    new BasicDevice(handle,
-                                    String::format("ATA Device {}-{}", handle, j),
-                                    hd_info.m_model_number,
-                                    hd_info.m_firmware_revision,
-                                    hd_info.m_serial_number,
-                                    DeviceType::MASS_STORAGE_DEVICE,
-                                    BasicDeviceID("ATA Device")));
-                LOGGER->debug("{} -> Drive{}, Partition{}: {} ({}): LBA {}-{}",
-                              c_logical_drive_id,
-                              i,
-                              j,
-                              partition->name,
-                              partition->type.to_string(),
-                              partition->start_lba,
-                              partition->end_lba);
-                _logical_drive_table[c_logical_drive_id++] = {.port_index            = (U8) i,
-                                                              .partition_table_index = (U8) j};
-
-                device_mapper(bus_device, device, nullptr);
-                _logical_drive_count++;
-            }
-        }
-    }
-
-    // ========================================================================================== //
-    // AHCI API
-    // ========================================================================================== //
-
-    auto AHCIDriver::get_discovered_hard_drives() -> LinkedList<HardDrive> {
-        LinkedList<HardDrive> hd;
-        for (size_t i = 0; i < PORT_LIMIT; i++) {
-            auto& p = _port_engine[i];
-            if (p.is_active()) hd.add_back(p.get_hard_drive_info());
-        }
-        return hd;
-    }
-
-    auto AHCIDriver::get_logical_drives() -> LinkedList<Partition> {
-        LinkedList<Partition> p_list;
-        for (size_t i = 0; i < _logical_drive_count; i++) {
-            LogicalDrive ld = _logical_drive_table[i];
-            p_list.add_back(*_port_engine[ld.port_index]
-                                 .get_hard_drive_info()
-                                 .partition_table[ld.partition_table_index]);
-        }
-        return p_list;
-    }
-
-    auto AHCIDriver::get_hard_drive_info(U8 hard_drive) -> HardDrive {
-        LogicalDrive ld = resolve_logical_drive(hard_drive);
-        if (ld.port_index == LogicalDrive::INVALID_PORT) return {};
-        U8 port_idx = ld.port_index;
-        if (!_port_engine[port_idx].is_active()) {
-            LOGGER->warn("No hard drive on port {} detected.", port_idx);
-            return {};
-        }
-        return _port_engine[port_idx].get_hard_drive_info();
-    }
-
-    auto AHCIDriver::send_ata_command(U8                     hard_drive,
-                                      void*                  buf,
-                                      size_t                 buf_size,
-                                      RegisterHost2DeviceFIS h2d_fis) -> size_t {
-        if (!_port_engine[hard_drive].is_active()) {
-            LOGGER->warn("Cannot send ATA command. No hard drive on port {} detected.", hard_drive);
-            return 0;
-        }
-        return _port_engine[hard_drive].send_ata_command(buf, buf_size, h2d_fis);
-    }
-
-    auto AHCIDriver::read(U8 hard_drive, void* buf, size_t buf_size, size_t lba) -> size_t {
-        LogicalDrive ld = resolve_logical_drive(hard_drive);
-        if (ld.port_index == LogicalDrive::INVALID_PORT) return 0;
-        U8 port_idx              = ld.port_index;
-        U8 partition_table_index = ld.partition_table_index;
-        if (!_port_engine[port_idx].is_active()) {
-            LOGGER->warn("Cannot read from device. No hard drive on port {} detected.", port_idx);
-            return 0;
-        }
-        Partition* p =
-            _port_engine[port_idx].get_hard_drive_info().partition_table[partition_table_index];
-        U64 t_lba = p->start_lba + lba;
-        if (t_lba > p->end_lba) {
-            LOGGER->warn(
-
-                "Cannot read from device. LBA not in partition range. Range: {}-{}, LBA: {}",
-                p->start_lba,
-                p->end_lba,
-                t_lba);
-            return 0;
-        }
-        return _port_engine[port_idx].read(buf, buf_size, t_lba);
-    }
-
-    auto AHCIDriver::write(U8 hard_drive, void* buf, size_t buf_size, size_t lba) -> size_t {
-        LogicalDrive ld = resolve_logical_drive(hard_drive);
-        if (ld.port_index == LogicalDrive::INVALID_PORT) return 0;
-        U8 port_idx              = ld.port_index;
-        U8 partition_table_index = ld.partition_table_index;
-        if (!_port_engine[port_idx].is_active()) {
-            LOGGER->warn("Cannot write to device. No hard drive on port {} detected.", port_idx);
-            return 0;
-        }
-        Partition* p =
-            _port_engine[port_idx].get_hard_drive_info().partition_table[partition_table_index];
-        U64 tLba = p->start_lba + lba;
-        if (tLba > p->end_lba) {
-            LOGGER->warn(
-
-                "Cannot write to device. LBA not in partition range. Range: {}-{}, LBA: {}",
-                p->start_lba,
-                p->end_lba,
-                tLba);
-            return 0;
-        }
-        return _port_engine[port_idx].write(buf, buf_size, tLba);
-    }
-
-} // namespace Rune::Device
