+++ Kernel/Src/Device/SConscript
@@ -21,13 +21,15 @@ sources = [
     build_env.File("ACPI/ACPICA/CLibFunctions.cpp"),
     build_env.File("ACPI/ACPICA/osl.c"),
     build_env.File("ACPI/ACPICA/osl-to-rune-bridge.cpp"),
-    build_env.File("AHCI/AHCI.cpp"),
-    build_env.File("AHCI/FIS.cpp"),
-    build_env.File("AHCI/GPT.cpp"),
-    build_env.File("AHCI/Port.cpp"),
-    build_env.File("AHCI/PortEngine.cpp"),
     build_env.File("Keyboard/Keyboard.cpp"),
     build_env.File("Keyboard/PS2Keyboard.cpp"),
+    build_env.File("MassStorage/AHCI/FIS.cpp"),
+    build_env.File("MassStorage/AHCI/GPT.cpp"),
+    build_env.File("MassStorage/AHCI/HostBusAdapterDriver.cpp"),
+    build_env.File("MassStorage/AHCI/Port.cpp"),
+    build_env.File("MassStorage/AHCI/PortDriver.cpp"),
+    build_env.File("MassStorage/AHCI/PortEngine.cpp"),
+    build_env.File("MassStorage/MassStorage.cpp"),
     build_env.File("PCI/ClassCode.cpp"),
     build_env.File("PCI/PCI.cpp"),
     build_env.File("PCI/Types.cpp"),
